Move `FeedErrorViewModel` to a separte file
