+++ EssentialFeedII/Feed Presentation/FeedPresenter.swift
@@ -29,18 +29,6 @@ public protocol FeedLoadingView {
     func display(_ viewModel: FeedLoadingViewModel)// happens in the uirefreshcontrol
 }
 
-public struct FeedErrorViewModel {
-    public let message: String?
-    
-    static var noError: FeedErrorViewModel {
-        return FeedErrorViewModel(message: nil)
-    }
-    
-    static func error(message: String) -> FeedErrorViewModel {
-        return FeedErrorViewModel(message: message)
-    }
-}
-
 public protocol FeedErrorView {
     func display(_ viewModel: FeedErrorViewModel)
 }
